Refactor Hdf5Loader into modular components and split interpolation tests
Refactor the monolithic Hdf5Loader.hpp into smaller, focused modules:
- Hdf5Types.hpp: Status enum, config struct, TableView and TableDevice types
- detail/Hdf5LoaderDetail.hpp: Low-level HDF5 reading utilities
- detail/Hdf5TableLoader.hpp: Table loading logic and validation
- Hdf5Loader.hpp: Umbrella header for backward compatibility

Split the large test_log_interpolate.cpp into focused test files:
- test_log_interpolate_basis.cpp: InterpBasis and IndexDelta tests
- test_log_interpolate_kernel.cpp: LogInterp kernel tests
- test_log_interpolate_deriv.cpp: Derivative basis function tests
- test_log_interpolate_2d.cpp: 2D interpolation tests
- test_log_interpolate_3d.cpp: 3D interpolation tests
- test_log_interpolate_4d5d.cpp: 4D and 5D interpolation tests
- test_log_interpolate_sweep.cpp: Sweep interpolation tests

Also refactored InterpLogTable.hpp and LogInterpolate.hpp to use the new
modular Hdf5Types.hpp for type definitions.
